컴벳 시스템 기본 골자 완성
이제 Bang 카드 완성되고 네트워크 적용해봐야함

+++ Assets/3.Script/CombatSystem/InGameEvent.cs
@@ -18,7 +18,7 @@ public abstract class InGameEvent
 
 public class CombatEvent : InGameEvent
 {
-    public float Damage { get; set; }
+    public int Damage { get; set; }
     public Vector3 HitPosition { get; set; }
     public Collider Collider { get; set; }
     public override EventType Type => EventType.Combat;
